Skip revalidation if reading a directly accessible address space
Since address translation is updated lazily, it need not be
correctly set up when reading from an address space that is directly
accessible by the underyling file format.

While this change is primarily an optimization (to prevent repeated
unnecessary revalidation), note that it may enable certain
initialization sequences that would otherwise fail on the
revalidation.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>