Set the dirty flag on cloned translation data
A cloned translation is uninitialized. Although users will usually
initialize it by setting an attribute that influences translation,
hence setting the dirty flag, they might use a value that does not
differ from the original. In that case, post_set hooks are skipped
and the translation is left unitialized.

For example, the test case uses the same root page table address to
trigger the bug.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>