Kill the concept of precious page I/O
The original intention was that higher-level page tables should not
be evicted when scanning lower-level page tables, although both
would appear to be referenced only once. However, there is
insufficient data on which pages should be considered precious, so
the "precious" field in struct page_io is always set to zero. At
this point, it's more candid to remove this field and all related
code.

NB if an in-flight cache entry is returned on a cache search, it is
still added to the precious list eventually.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>