Add low-overhead cache to read64/read32 kdumpfile callbacks
It turns out that the overhead of searching the page cache is not so
small, especially when reading a very small amount of data, such as
a single 32-bit or 64-bit value.

This cache must add as little overhead as possible, e.g. it must not
do any locking. That's why it is added to each kdumpfile context and
its size is limited.

The cache replacement algorithm is primitive. Essentially, it keeps
around READ_CACHE_SLOTS buffers of size READ_CACHE_SIZE, evicting
the least recently added slot when necessary. This may cause an
immediate refill of a slot that has just been evicted, but this
issue is offset by extremely low cache overhead (and the fact that
the underlying data is already cached, just with more overhead).

Note that the cache elements are bigger chunks of memory than a
single value. This is useful with sequential memory scanning, as it
effectively performs readahead.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>