Remove mutex from the cache object
Each cache currently contains a mutex to allow concurrent access
from multiple threads. This works well, but it imposes the locking
overhead even on users who know that there cannot be any race.

Additionally, there are two kdump caches (the file cache and the
page cache), but one lock would be sufficient for both.

Finally, the lock is not needed if the shared data is locked for
writing (e.g. the format probe functions).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>