Optimize get_page for directly supported address spaces
For library users, the most critical path are reads using the
underlying dump file's native address space, because such calls
cannot be optimized any further. Give some hints to the compiler
that the non-translated part should be inlined (and the translation
can be in a separate function). It can save the call overhead in
read_locked(). Callgrind shows a speed-up of 10 Ir per call.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>