Optimize highest_linear() without machphys->kphys on each page
The routine is used to find the extent of various linear mappings.
In reality, there is always at least one gap page after such
mappings. The only exception is a full 512M Linux kernel text
mapping immediately followed by a kernel module. This case requires
special handling at the ktext/module boundary:

  - either modules start at this address, or
  - a ktext mapping continues at this boundary.

The latter case is possible only with a kASLR-enabled kernel, but in
that case at least one gap page is inserted. Actually, this is
currently not needed at all, because modules cannot be loaded below
0xffffffffc0000000 in a kASLR kernel, but this limitation may be
easily changed, so let's be prepared...

Signed-off-by: Petr Tesarik <ptesarik@suse.com>