ELF: Make sorted lists of LOAD segments
The search can be simplified by using a sorted list. The search is
still sequential, so not suited for a very large number of segments,
but combined with the "remeber last found" feature, it actually
performs pretty good for the common case, and also exhibits good
results with my proof-of-concept makedumpfile rewrite.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>