python: Fix "dictionary changed size during iteration" error in python3
In python2, the dict.values() method returns a list that is a snapshot, so
the dictionary object can be modified while iterating over the list.  In
python3, the same method returns a view object, and the dictionary object
cannot be modified while iterating over the view object.

Fixed by using list(dict.values()) for python3.

[ An alternative solution could be to:
- move the corresponding code into a function _here, changing the scope of
  the variables to local, which makes sure globals() is not
  changed during iteration,
- call _here, and
- call del _here. ]