Update version_code attributes using revalidate hook
The release string and/or xen major/minor version attributes may be
changed from an arbitrary place. Instead of updating the dependent
attributes explicitly, add a revalidate hook to make sure that the
version_code value is correct when requested.

I considered updating the value from the post_set hook, but:

 1. The Xen version code cannot be updated atomically (major and
    minor are two attributes), producing an invalid version_code
    temporarily.
 2. The Linux release string may not even be valid; this condition
    would either have to be ignored completely, or the setting of an
    unparseable release string would always fail.

It is still necessary to implement post_set hooks to make sure that
the version_code attribute is not left unset.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>