Do not fail if Xen extra version string is not stored in the dump
The extra version is updated when OS type is set, so if the dump
does not contain the extra version (e.g. it is incomplete), it is
not possible to initialize address translation for it, although it
is not even used for that purpose.

Ignore missing data for extra version (but make sure that other
errors, such as I/O errors are still reported).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>