x86_64: Convert late_init to post_addrxlat
This fixes the elf-dom0-no-phys_base test case. The root cause is
that dumpdata first sets addrxlat.ostype and then file.fd, so the
ostype post_set hook is run when architecture is unknown. The
fallback code in init_linux_phys_base() is not run later.

The real intention of the hook is to make sure that phys_base is
known after address translation is initialized, so it is best
implemented as a post_addrxlat hook.

Note that vtop_init() is called recursively when phys_base is set to
the fallback value.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>