Live source: Detect Xen presence and domain type
Xen is normally detected by its specific file format or specific ELF
notes. When debugging a running system, none of that is available,
so live memory source works only on bare metal.

I'm not using longjmp() to avoid a global variable. Instead, the
signal handler manipulates the instruction pointer as needed to let
the caller of xen_cpuid() know that the ud2a instruction was not
emulated.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>