python: fix use-after-free with addrxlat.Context callbacks
We keep a pointer to the Context object as the callback data pointer
but don't take a reference to it.  Later, after the Context object
has all references dropped from within Python and is garbage collected,
the callback can fire and we get a use-after-free.

This commit fixes it by taking a reference in install_cb_hook and
releasing it in cb_hook.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>