Fix `make check` when configured with `--with-pic`
If libkdumpfile is configured with `--with-pic`, `make check` fails
with:

  /bin/ld: test-fcache.o: in function `main':
  /home/osandov/repos/libkdumpfile/src/kdumpfile/test-fcache.c:419: undefined reference to `_kdumpfile_priv_fcache_new'
  /bin/ld: test-fcache.o: in function `test_basic':
  /home/osandov/repos/libkdumpfile/src/kdumpfile/test-fcache.c:111: undefined reference to `_kdumpfile_priv_fcache_get'
  /bin/ld: /home/osandov/repos/libkdumpfile/src/kdumpfile/test-fcache.c:132: undefined reference to `_kdumpfile_priv_fcache_get'
  ...

This is because the static library is compiled with -DPIC but the test
binary is compiled without it, so they use different names for internal
functions. Fix it by always using the private naming; even though it's
unnecessary for shared libraries, it doesn't hurt either.