Convert attribute hash list to a hlist
Use a double linked list to store attributes. This allows to add
new fields to struct attr_data without too much impact on the
memory footprint. Of course, more metadata is now needed to keep
track of allocated blocks in the C library. However, on a couple of
sample dump files, total memory usage decreased by 1.5 % and 3.8 %,
respectively. The only cost seems to be the number of calloc calls,
which increased by approx. 50 %, but the assumption is that:

  1. The implementation of calloc() is reasonably fast.
  2. Attribute hash tables are populated only once at startup.

Note that newly allocated attributes must not be cleared in
new_attr(), because that would clear the list node
metadata. Instead, is zero-initialized by allocating it with
calloc().

Another caveat is walking attributes while deallocating them. This
used to be safe, because a "free" attribute was merely marked by
settung its parent field to NULL. Since attribute data is now truly
freed, callers must avoid dereferencing an attribute after calling
dealloc_attr() on it.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>