Use NULL as root attribute's parent node
The root node currently points back to itself. This was needed,
because the NULL value was already used to mark free entries. Since
free entries are deallocated immediately, they cannot appear in any
lookup, so NULL value is never used and can be repurposed to mark
end of parent chain.

As a side effect, this change fixes a subtle bug that an attribute
path could be prefixed by any number of dots (that's the added
conditional in keycmp, which is now needed to prevent NULL-pointer
dereferences).

Also, link_attr() broke badly, but since there's only one caller,
I've merged it into new_attr().

Signed-off-by: Petr Tesarik <ptesarik@suse.com>