Fix `make check` when configured with `--with-pic`
The basic idea of having internal names was to see the source code
names in a debugger. Unfortunately, libtool is brain-dead, so there
is no way to distinguish whether a symbol can be hidden from other
users.

This is a modification of the change requested in PR 24. Sanitized
names will be used by default for shared libraries and static
libraries alike, but unmodified names can be still turned on by
running `configure` with `--enable-debug`.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>