Update the PRSTATUS attribute if register value is modified
It is OK to set the attribute value for cpu.<num>.reg.<reg>. But
such changes are not reflected in the PRSTATUS blob attribute.
Add a post-set hook for that purpose.

Note that the hook does not update anything if the new value is
flagged as invalid. If you remove this check, setting the
initial (invalid) value for all register attributes will zero out
PRSTATUS.

Oh, and add a test for the new behaviour.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>