Convert attribute template parent to a union
This field is used only by global attribute initialization and
overrides. However, the use of pointers means that the global
initialization array adds unnecessary relocations, because the
code translates each address back to a global attribute index.

This can be solved by converting to a union and using designated
initializers in the global array.

I have chosen to keep the old type first, so non-designated
initializers need not be adjusted, reducing churn.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>