Improve ppc64 register names
A few improvements on the register naming:

- gpr00..gpr31 can be called simply r0..r31
- pc (Program Counter) is a more common name than nip (Next
  Instruction Pointer)
- mq was used only on 601; this slot is used for softe on PPC64.
- trap is missing; all following registers are off by one
- the first register after dsisr is the result, reserved for the
  result of the currently running system call
- rx1..rx7: no idea where this comes from, but since these registers
  are not used in a kernel crash dump anyway, they can be ignored

ELF_NGREG is 48, not 49. The extra space is used for pr_fpvalid (and
a 4-byte hole at the end of the structure). However, on crash dump,
PPC_ELF_CORE_COPY_REGS copies registers from struct pt_regs, and
there are only 44 registers in struct pt_regs. Rest of the PRSTATUS
structure is always filled with zeroes.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>