xc_core_dump: Fix single PFN lookup
Part of the loop condition was reversed. Since the singles array is
sorted in ascending order, lookup should continue as long as the
searched PFN is _greater_ than the current one.

The current code fails if there is more than one PFN in the
singles array and other than the first PFN is searched.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>