Handle compound types attr_has_value()
According to documentation, hooks should be skipped if attribute
value is not changed. This is not the case for bitmaps and blobs.

Let's assume that two different instances of the same value are not
the same value. Although we could check the underlying buffers, the
objects themselves are mutable, so there is no guarantee that they
do not diverge in the future, and then the hooks would never be
called. Apart from that, comparing bitmap data can be
time-consuming.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>