addrxlat: Add a get_page/put_page callback API
The current API to read data from a dump memory mostly works, but
it has limitations:

- The callback does not get enough information about the intended
- use of the requested data to make better caching decisions.
- The call overhead is not negligible when reading a complete
  page table sequentially.
- Some data structures require a specific byte order, regardless of
  the primary byte order of the dump file.

The change requires adding byte order knowledge to addrxlat. I have
decided to add it to the buffer metadata, because it replaces the
logic of read32/read64 callbacks. Since it does not change across
calls to get_page(), it could probably be added more
permanently (e.g. as a parameter at initialization), but at least
one user (libkdumpfile) does not know the information at that time,
so let's keep it in the buffer metadata for now.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>