Convert C test suite to use the new get_page API
The implementation is trivial. For example, put_buffer() is not
even needed here, because these stub buffers are not managed
dynamically.

The only catch is that tests/addrxlat.c does not know if the entity
has 32 or 64 bits. That's why I'm forcing the use of little endian
as storage format---an integer stored in a 64-bit field that is
dereferenced as a 32-bit memory cell will still yield correct
32-bit value.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>