python: Convert addrxlat callbacks to get_page
I didn't want to introduce a separate Python type to represent
addrxlat_buffer_t, so the API is a bit rough but hopefully still
acceptable.

The callback function now returns a tuple. The first element is an
object that implements the buffer protocol. The (optional) second
element is the byte order. The address is passed as argument and
should be adjusted in place.

The test case must return actual bytes with a well-defined
endianity. I've chosen big endian, because you needn't read the
values backwards. And it also serves as a test case for big endian
buffers, which are less common these days.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>