addrxlat: Implement a trivial read cache
This is a rough replacement for the caching mechanism that was
removed from libkdumpfile in commit c1993d5983.

I'm using the same tiny cache size, as it has proven sufficiently
good in practice. It may be a good idea to review this magic
constant when 5-level paging is implemented.

Note that addrxlat now holds a reference to all cached pages. This
allows zero-copy acces to the page data, but it also increases the
utilization of the kdumpfile cache, because all cache slots will be
used eventually.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>