tests: check empty ELF dumps for IBM POWER
This architecture requires a modification to the elf-empty script,
because it does not have any default page size (would it be 64k?
4k? 16k?).

Note that the page size is left uninitialized on an empty dump
file. That's why cache was not allocated, and that was the
motivation for the previous commit.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>