Fix a crash if vtop_init() is called without an architecture name
Calling addrxlat_sys_os_init() requires a non-NULL architecture
name. However, vtop_init() is now called from revalidate_xlat(),
and that can be triggered by a user calling kdump_get_addrxlat().

If architecture is unknown, address translation cannot be improved
beyond the provisional dummy system created at startup. This is in
fact not an error, so keep the current state and claim success.

The dirty flag is not removed. If the architecture is set later,
translation will be set up automatically as needed.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>