Handle ELF dumps with lots of sections or segments
ELF files with >= 0xff00 sections have an e_shnum of 0, non-zero
e_shoff, and the real number of sections in the sh_size of of the first
section header. Similarly, files with >= 0xffff segments have an e_phnum
of PN_XNUM (0xffff) and the real number of segments in the sh_info of
the first section header. See elf(5).

I haven't seen the former in practice, but the latter is easily hit on
ELF dumps of large systems produced by makedumpfile. Make sure to handle
both in init_elf{32,64}.

Signed-off-by: Omar Sandoval <osandov@fb.com>