Use Python distutils to build and install Python bindings
Although GNU Automake supports building Python modules and
extensions, this approach does not allow to use standard Python
packaging methods.

One of the reasons to stay with GNU Automake has been lack of
support for GNU Libtool, which is used to build the main library.
In particular, linking to a non-installed library requires
libtool. This commit adds a `libtoolize` module, which modifies
the default compiler to use libtool.

Some parameters must be passed from GNU autotools to distutils.
This is solved by generating a `setup.cfg` file from Makefile.

There are a few gotchas:

1. To build a shared library, libtool requires the `-rpath` option
   at link time, but Python distutils do not care about the
   destination directory during build. I have to pass `pyexecdir`
   explicitly to the build command.

2. The build tree may be different from the source tree (aka VPATH
   build). The source directory is passed down via a `srcdir`
   option and used as `package_dir`. It must be also explicitly
   prepended to all extension source files.

3. The libtool files for libkdumpfile and libaddrxlat must be
   added as `extra_objects`, because the libraries are not yet
   installed. Full path to the objects must be provided, and that
   one is relative to the build tree, so we also pass down
   `top_builddir` as an option.

Note that tests have not been converted (yet), because they should
not be installed, but distutils do not seem to implement that
functionality.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>