Add PTE mask page table translation parameter
AMD EPYC processors introduce an optional SME bit in the page
tables to indicate encryption. The location of this bit is stored
in bits 0:5 of CPUID Fn8000_001F[EBX], i.e. it is not known at
build time and must be configurable through a parameter.

There might be other good use cases for this mask, because the API
is quite generic...

Signed-off-by: Petr Tesarik <ptesarik@suse.com>