Convert vmcoreinfo.raw attributes to blob type
The string type are intended for relatively short strings that are
part of the attribute. This does not really fit the vmcoreinfo raw
value, but no better type was available at that time.

This has worked out quite well (mostly because few users are using
this attribute), but there are some subtle issues. For example,
the raw value must be NUL-terminated; the NUL byte is added
artificially, but it is not possible to know if NUL was actually
part of the value, or not. This is not so important for parsing
the VMCOREINFO content, but if the raw data should be processed in
other ways (e.g. saved to another file), we'd better provide an
interface that is binary clean.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>