Fix attribute value reference counting
The attribute handling code does not properly deal with reference
counts. Part of the issue is that object ownership has never been
clarified, so it's not clear what should happen to an object which
is passed as the new value argument to set_attr().

Since set_attr() may fail in post_set hook, at which point the new
value is already set and the change is not rolled back, the only
sane thing is to claim ownership of the object in any case.

But if the attribute data owns a refcounted object, then the
reference must be dropped when the attribute data is freed or
changed to another value. For dynamically allocated strings, the
code is duplicated in a few places. Move the code to a single
place - discard_value(), even if it means faking an attribute that
has not been set.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>