Fix kdump_vmcoreinfo_raw()
The conversion of raw vmcoreinfo attributes to a blob type
requires change of kdump_vmcoreinfo_raw(), as reported by
@sdimitro in issue #32.

This function has always been flawed, because the lifetime of the
return value is underspecified. For example, the pointer is
invalidated by any change to the vmcoreinfo attributes. The call
is inherently racy in a multi-threaded program; in an extreme
case, the pointer may become invalid before it is even returned to
the caller.

The above issue can be fixed by returning a dynamically allocated
copy of the value. However, such a fix also requires changing the
callers to avoid memory leaks. Luckily, the type of the parameter
must change, so compilers can issue a warning. so the breakage is
not completely silent.

Fixes: b4b3c94
Signed-off-by: Petr Tesarik <ptesarik@suse.com>