Fix search for highest mapped address
When the beginning of a page table is reached, the lower-level
table indices are initialized to the last element index of the
current table. That's wrong. It must be initialized to the last
element index of the corresponding lower-level table.

Note: This code is currently unused, but it will be used and tested
for AArch64.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>