aarch64: Simplify provisional address translation in get_linux_pgtroot()
The code apparently tries to mimic Linux __pa() macro, but it's
unnecessarily complex. It is used to get the physical address of
swapper_pg_dir, which is known to be a kernel text symbol, so there
is no need to check whether it is an address in the linear
mapping.

Furthermore, if kimage_voffset is unknown, the whole fallback logic
can produce a bogus physical address. The only reason it never
actually does is because both kimage_voffset and PHYS_OFFSET were
added with the same Linux commit 20a166243328c14a, so if there is no
kimage_voffset, there is no PHYS_OFFSET either, and initialization
fails.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>