aarch64: Set up the linear mapping if possible
Although the linear mapping region is indeed covered by page
tables, the main benefit of having one is that a _reverse_ direct
mapping can be created. This allows to perform physical-to-virtual
translation, which is required if the read callback does not handle
physical addresses.

However, if the linear mapping cannot be created, the error is
ignored, because it is (usually) non-fatal.

Most of the commit extends the test suite to add data from three
sample dumps of a QEMU VM running various kernels.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>