aarch64: Support non-4K page sizes
Instead of hard-coding the paging form for 4K pages, calculate the
fields from the VA width in bits and page size.

The LSB field is always equal to page shift. That's calculated from
the page size with the ffsl() function. On AArch64, page size is
always equal to the maximum size of a single translation table (see
Memory translation granule size in the Reference Manual). Since
entries in that table are 64 bit (8 bytes = 2^3 bytes) long, the
maximum size of the corresponding address field is page_shift - 3.

The total number of bits must not be greater than VA_BITS, i.e. the
highest-level translation table may be smaller.

There is no default page size, i.e. pagesize must be explicitly
included in addrxlat options. That's why the unit tests had to be
modified.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>