aarch64: Do not open-code ADDR_MASK() in map_linux_aarch64()
There is already a macro that does exactly the same thing.

I'm also removing the va_range variable. GCC is clever enough to
find the common subexpression. However, the real reason is that
the lower VA range may potentially be of a different size than the
upper VA range, although this is not implemented.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>