tests/multixlat: Use the new addrxlat option attributes
This can be used as an example for converting existing code that
uses addrxlat.opts.

The old interface sets the KDUMP_ATTR_XLAT_OPTS_POST attribute to
rootpgt=MACHPHYSADDR:0x_address_ (as a string).

The new interface sets:

- KDUMP_ATTR_XLAT_FORCE ".rootpgt.as" = ADDRXLAT_MACHPHYSADDR
- KDUMP_ATTR_XLAT_FORCE ".rootpgt.addr" = _address_

The new code is a bit more verbose (at least for a full address
attribute), but the information is better structured (and we don't
have to print into a temporary buffer).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>