Clone subtrees more efficiently
Although the current code works, a complete attribute path is
constructed for each cloned child, then the closest cloned parent
is searched and then the attribute is cloned.

This can be done much easier, because it is a known fact that the
target directory attribute has just been cloned, so we can directly
use clone_attr().

Signed-off-by: Petr Tesarik <ptesarik@suse.com>