Pass addrxlat options in a parsed form
It is somewhat silly that callers of addrxlat_sys_os_init() now
create a string option from individual components, which gets
immediately parsed again in libaddrxlat.

Change the API to accept parsed option values instead.

This is a backwards-incompatible change, and first such change
since the release of v0.4.1, so bump version info.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>