Simplify addrxlat_sys_os_init() usage
Users of this API are still required to pass a pointer to
addrxlat_osdesc_t.  However, this structure now merely contains a
pointer to the option array and the number of options, which could
be more easily passed as two arguments.

It is no coincidence that the new function signature bears more
resemblance to the main() function. ;-)

Signed-off-by: Petr Tesarik <ptesarik@suse.com>