Rewrite ostype_attr() without global index map
The ostype_attr() function is a fast hack. It would be justified in
case performance was an issue, but these dictionary lookups are not
in the critical path, so let's do it the "hard" way, looking up
each level of indirection using standard functions.

The benefit is that the code itself now ensures that all
OS-type-specific attributes are indeed found under $ostype.$name,
making it harder to introduce silly copy&paste mistakes.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>