addrxlat: Use a string for the os_type option
Let's use a string for the name of the operating system, rather
than an enum.

There is no real support for enums in the option interface. Adding
one just for this one option would be overkill. As a result, the
option is implemented as a number, which is harder to maintain than
a string representation.

Plus, the architecture is already passed as a string, so the API is
a bit more consistent now.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>