addrxlat: Add a phys_bits option
The number of physical address bits is variable on some platform.
It is a better way to describe the difference between PAE and
non-PAE ia32 paging. The physical address range is the root cause;
halving the number of translation table entries each level of
paging and the need to add one more level of paging is merely a
consequence.

As a bonus, this option can be later reused for AArch64 to convey
the state of 52-bit (I)PA.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>