addrxlat: Add a virt_bits option
The number of virtual bits must be configurable on some platforms.
So far, paging levels have been used for that purpose, but that's
insufficient for AArch64, where the hardware allows to set an
arbitrary virtual address size in bits (cf. TxSZ field in the
corresponding (V)TCR_ELx register), possibly reducing the size of
the top-level translation table.

Use virt_bits instead of levels where the size of the virtual
address matters.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>