x86_64: Use virt_bits=48 if VMCOREINFO lacks pgtable_l5_enabled
If pgtable_l5_enabled is not found, the it is either missing from
the VMCOREINFO, or there is no VMCOREINFO at all. In the former
case, it is most likely an old kernel without support for 5-level
paging, and we may assume 48 bits for the virtual address.

Unfortunately, checking for the existence of VMCOREINFO is not
straightforward. We pick a symbol that has always been
included (_stext), and if it is not found, we assume that
VMCOREINFO is not available.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>