aarch64: Always set up the hardware translation map
This translation is defined by hardware. The code already makes
sure that the relevant parameters are available, so the hardware
map can be created even if the OS type is unknown.

This can also be the default KV_PHYS translation map. With that,
address translation is sufficiently initialized, so the code should
no longer fail because of unknown OS type.

Oh, and add a test case, of course.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>