tests: Fix symbol storage in xlat-os
The symbol value always occupies the whole buffer passed to the
write_page callback. The only difference between little endian and
big endian is whether the buffer should be copied to the beginning
of the value (little endian) or end of the value (big endian).

Besides, the size of the value is clamped to the target size and
stored in a local variable sz, but the full buffer size is copied,
potentially causing buffer overflow on the target value variable.

Fixes: 1c1b359688e1a97bece751e682aca6ed6928361d
Signed-off-by: Petr Tesarik <ptesarik@suse.com>