Fix handling of 64-bit addresses on 32-bit platforms
The get_page_size() inline function returns a size_t. This type is
big enough for all supported page sizes, even on a 32-bit host,
but it must be cast to a kdump_addr_t to get a page mask.

I'm not changing the return type of get_page_size(), because using
64-bit arithmetics would incur some performance penalty (even on
many 64-bit systems).

References: #40
Signed-off-by: Petr Tesarik <ptesarik@suse.com>