fcache: Promote pgsz and mmapsz to off_t for masking off low bits
On 32-bit systems, size_t is a 32-bit integer, whereas off_t can
be a 64-bit integer (when built with -D_FILE_OFFSET_BITS=64). To
mask off the low bits, the high bits must be ones, i.e. the size_t
must be cast to an off_t before doing the logical NOT.

References: #40
Fixes: 0da13148b4579ffee5a4d0fac1680bc4b26c92ef
Signed-off-by: Petr Tesarik <ptesarik@suse.com>