Fix libkdumpfile unit tests if static libraries are not built
The unit tests check parts of the internal API. These functions
are not exported in a dynamic link. Passing `-static` to the
linker does not help if there is no static library (e.g. when the
project was configured with `--disable-static`).

Make a convenience library (libcheck) from the same object files
as libkdumpfile. This library is always linked statically, which
ensures that the internal API is exposed to the test suite.

References: #42
Signed-off-by: Petr Tesarik <ptesarik@suse.com>