Clean up pkg-config files
Libraries from packages that support pkg-config should not be listed
again under Libs.private. For the compression libraries, lzo2 and
snappy do not currently use pkg-config. If any of them starts using
pkg-config, we should prefer that method, but we should still
support older versions without pkg-config.

Because of the above, the pkg-config file cannot simply use the
linker flags. They should be added only if there is no corresponding
pkg-config requires. That's what the newly introduced *_PC_LIBS
variables are good for.

Both Requires.private and Libs.private lines can be removed if they
are not needed.

References: #42
Signed-off-by: Petr Tesarik <ptesarik@suse.com>