Track number of dump files pending open
There can be more than one file descriptor, but the attribute API
does not provide a method to set them atomically. Instead of
complicating the API, track the number of files without a file
descriptor and open the dump when all file descriptors are set.

The file.fd attribute is left for compatibility, but it is now
implemented using fdset. Setting file.fd = X translates to

  - clear file.fdset
  - file.fdset.number = 1
  - file.fdset.0 = X

Signed-off-by: Petr Tesarik <ptesarik@suse.com>