Do not try to set file.fd when using the file.fdset API
Setting the file.fd attribute emulates setting fdset.number to one
and fdset.0 to the file.fd value. Setting fdset.0 tries to make
sure that file.fd is also set, but that's broken, because it
resets fdset.number to one and then recursively sets fdset.0.

It would be quite complex to achieve proper compatibility. More
importantly, it's not worth the effort, because there are probably
no users relying on it. So, instead, keep file.fd only if it was
actually used to set the dump file descriptor and clear it as soon
as the fdset API is used in an incompatible way.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>