tests: Fix mkdiskdump for split dump files
The organization of data in a split dump file is weird...

The _complete_ bitmap is copied to every split dump file, but then
only the page descriptors within <start_pfn; end_pfn) are included
in the file.

The mkdiskdump utility cannot create such files, so it's not
possible to write a test case (and thus to add support for split
files).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>