Allow any order of split dump files
Originally, the dump files were assumed to be sorted by ascending
PFN. However, the results can be confusing if split files are
specified in a different order. Instead of improving error
detection, allow to specify the files in any order and sort them
when opening the file set.

This means that the index in the pdmap array no longer corresponds
to the file index under the file.fdset attribute, so this other
index must be stored in the page descriptor map.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>