python: add kdumpfile_object_from_native to create object from open ctx
When using libdumpfile in a project that uses the library directly but
also has Python extensions, it's useful to be able to create the Python
object from the existing object rather than creating a new context that
won't share caches.

This adds a new kdumpfile_object_from_native exported function that can
be used by such projects by dlopening the already-loaded Python module.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>