Replace ffs() with ctz()
No big difference, but ctz() should be a tiny bit more efficient
than ffs(), as it does not have to deal with zero argument. The
check for zero is either already performed as part of the loop
condition, or adding it allows to remove another check on the
result of ffs(), so either way it does not add any overhead.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>