addrxlat: Fix put_page() callback design
The callbacks are designed so that they can be chained (thanks to
the update hook). This mechanism does not work well for the
put-page callback, however, because the callback function cannot
easily determine whether the buffer was allocated by the
corresponding get-page callback or by another get-page callback
further down in the chain. In other words, it is difficult for
put-page callback implementors to decide whether a given buffer
should be handled or passed down to the next chained put-page
callback.

As a matter of fact, all implementations included in the source
tree are broken in this regard. AFAICS the only way to implement a
get-page/put-page callback pair properly would be to track all
buffers that have been allocated and then check the address of the
buffer itself. That's just too complicated...

Instead, let the get-page callback store the corresponding
put-page pointer in each buffer metadata. This works even if page
buffers are allocated by different get-page implementations.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>