addrxlat: Pass the address of the callback struct itself to callbacks
Rename the `data` field to `priv` and pass the address of the
callback structure itself to the callback functions instead of just
the data pointer. The callbacks can still access the `priv` field
through this pointer, but this changed API is better for defining
internal callbacks that should not depend on this user private
field.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>