addrxlat: Redesign the callback API
Get rid of the cb_hook callback. This seemed like a good idea when
it was added, as it provided a way to chain callbacks without
breaking source-code compatibility with programs that had been
written before that feature, but it turns out the concept is
hard to grasp and broken in various subtle ways.

Since commit 90d5f9b4f0882d70bdea4b1053e74a9c5eb9fbd6 already
broke source-code compatibility, let's use the opportunity and
also fix the callback API.

The main change is that the callback structure is now allocated
and initialized by libaddrxlat. This should allow to extend the
callback structure without breaking the library ABI.

Other changes:

- There is no longer an "original callback" and "effective
  callback".
- Callback implementors do not have to keep a copy of the old
  callbacks to call the previously used callback function. Simply
  use the `next` field.
- To remove a callback implementation anywhere in the chain, call
  `addrxlat_ctx_del_cb`, which takes care of the rest.
- All callbacks are automatically cleaned up when the addrxlat
  context is freed.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>